Fixing corruption writing out vector<string> to silo.

+++ src/PBGWraps/Silo/SiloModule.py
@@ -370,7 +370,7 @@ class Silo:
         x.add_constructor([param("int", "maxopts", default_value="1024")])
 
         # Methods.
-        for ValueType in ("double", "int", "std::string"):
+        for ValueType in ("int", "double", "std::string"):
             ValueBase = ValueType.split(":")[-1]
             x.add_method("addOption", "int", [param("int", "option"), param(ValueType, "value")],
                          template_parameters = [ValueType],
